Transform BatchingHttpEngine into an HttpInterceptor
