+++ apollo-runtime/src/commonMain/kotlin/com/apollographql/apollo3/ApolloClient.kt
@@ -24,9 +24,11 @@ import com.apollographql.apollo3.interceptor.NetworkInterceptor
 import com.apollographql.apollo3.internal.defaultDispatcher
 import com.apollographql.apollo3.mpp.assertMainThreadOnNative
 import com.apollographql.apollo3.network.NetworkTransport
+import com.apollographql.apollo3.network.http.BatchingHttpInterceptor
 import com.apollographql.apollo3.network.http.HttpEngine
 import com.apollographql.apollo3.network.http.HttpInterceptor
 import com.apollographql.apollo3.network.http.HttpNetworkTransport
+import com.apollographql.apollo3.network.http.canBeBatched
 import com.apollographql.apollo3.network.ws.WebSocketEngine
 import com.apollographql.apollo3.network.ws.WebSocketNetworkTransport
 import com.apollographql.apollo3.network.ws.WsProtocol
@@ -372,6 +374,28 @@ private constructor(
           )
       )
       enableAutoPersistedQueries(enableByDefault)
+    }
+
+    /**
+     * Batch HTTP queries to execute multiple at once.
+     * This reduces the number of HTTP round trips at the price of increased latency as
+     * every request in the batch is now as slow as the slowest one.
+     * Some servers might have a per-HTTP-call cache making it faster to resolve 1 big array
+     * of n queries compared to resolving the n queries separately.
+     *
+     * @param batchIntervalMillis the interval between two batches
+     * @param maxBatchSize always send the batch when this threshold is reached
+     *
+     * See also [BatchingHttpInterceptor]
+     */
+    @JvmOverloads
+    fun batching(
+        batchIntervalMillis: Long = 10,
+        maxBatchSize: Int = 10,
+        enableByDefault: Boolean = true,
+    ) = apply {
+      addHttpInterceptor(BatchingHttpInterceptor(batchIntervalMillis, maxBatchSize))
+      canBeBatched(enableByDefault)
     }
 
     @Deprecated("Used for backward compatibility with 2.x", ReplaceWith("httpMethod(HttpMethod.Get)", "com.apollographql.apollo3.api.http.httpMethod", "com.apollographql.apollo3.api.http.HttpMethod"))
