+++ apollo-runtime/src/commonMain/kotlin/com/apollographql/apollo3/network/http/HttpInterceptor.kt
@@ -9,6 +9,7 @@ interface HttpInterceptorChain {
 
 interface HttpInterceptor {
   suspend fun intercept(request: HttpRequest, chain: HttpInterceptorChain): HttpResponse
+  fun dispose() {}
 }
 
 internal class DefaultHttpInterceptorChain(
