+++ apollo-runtime/src/commonMain/kotlin/com/apollographql/apollo3/network/http/HttpNetworkTransport.kt
@@ -103,6 +103,7 @@ private constructor(
   }
 
   override fun dispose() {
+    interceptors.forEach { it.dispose() }
     engine.dispose()
   }
